Rename & move gRPC utilities
